Create reusable overlay and mobile sidebar components
